+++ src/Header/WWWAuthenticate.php
@@ -3,26 +3,43 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
-use Innmind\Http\Header;
+use Innmind\Http\{
+    Header,
+    Header\WWWAuthenticate\Challenge,
+};
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
  */
 final class WWWAuthenticate implements Custom
 {
-    private Header $header;
+    /**
+     * @param Sequence<Challenge> $challenges
+     */
+    private function __construct(
+        private Sequence $challenges,
+    ) {
+    }
 
     /**
+     * @psalm-pure
      * @no-named-arguments
      */
-    public function __construct(WWWAuthenticateValue ...$values)
+    public static function of(Challenge ...$challenges): self
     {
-        $this->header = new Header('WWW-Authenticate', ...$values);
+        return new self(Sequence::of(...$challenges));
     }
 
     #[\Override]
     public function normalize(): Header
     {
-        return $this->header;
+        return new Header(
+            'WWW-Authenticate',
+            ...$this
+                ->challenges
+                ->map(static fn($challenge) => new Value\Value($challenge->toString()))
+                ->toList(),
+        );
     }
 }